pyproject.toml
[project]
name = "onvif-camera-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "onvif-zeep-async>=4.0.4",
    "python-dotenv>=1.2.1",
]
